# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 経理・会計オールインワンAIアプリケーション API
  version: "1.0.0"
  description: |
    中小企業〜中堅企業向け「経理・会計をこのアプリ1つで完結させる」オールインワンAIプロダクトのAPI仕様。

    対応する統合DB設計は `docs/03_database_design.md` / `sql/001_initial_schema_all_in_one.sql` を参照。

    ## 設計原則
    - AIは提案(`/ai/suggestions`)のみを行い、確定処理(仕訳確定・税額計算等)は本APIの決定的ロジックが行う。
    - 状態遷移・業務アクションはサブリソース(例: `POST /journal-entries/{id}/post`)として明示する。
    - 全エンドポイントはマルチテナントであり、`X-Tenant-ID` ヘッダーとJWTクレームの双方で検証される。
    - 副作用を伴うPOST操作は `Idempotency-Key` ヘッダーによる冪等性をサポートする。
  contact:
    name: 経理会計オールインワンソフト開発チーム
servers:
  - url: https://api.keiri-kaikei.example.com/v1
    description: 本番環境
  - url: https://api.staging.keiri-kaikei.example.com/v1
    description: ステージング環境
  - url: http://localhost:3000/v1
    description: ローカル開発環境
...
security:
  - BearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: 勘定科目詳細を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 勘定科目詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Accounts]
      summary: 勘定科目を更新する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountUpdate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /departments:
    get:
      tags: [Departments]
      summary: 部門一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: 部門一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentListResponse" }
    post:
      tags: [Departments]
      summary: 部門を新規作成する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DepartmentCreate" }
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentResponse" }
  /departments/{id}:
    patch:
      tags: [Departments]
      summary: 部門を更新する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DepartmentCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentResponse" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ----------------------------------------------------------------------------
  # Journal Entries (会計コア)
  # ----------------------------------------------------------------------------
...
  /bank-accounts:
    get:
      tags: [BankAccounts]
      summary: 銀行口座一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: 銀行口座一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BankAccountListResponse" }
    post:
      tags: [BankAccounts]
      summary: 銀行口座を登録する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankAccountCreate" }
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BankAccountResponse" }
  /bank-accounts/{id}:
    get:
      tags: [BankAccounts]
      summary: 銀行口座詳細を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 銀行口座詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BankAccountResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [BankAccounts]
      summary: 銀行口座情報を更新する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BankAccountCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BankAccountResponse" }
...
  /customers:
    get:
      tags: [Customers]
      summary: 顧客一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - name: q
          in: query
          description: 名称・カナ名のあいまい検索
          schema: { type: string }
      responses:
        "200":
          description: 顧客一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerListResponse" }
    post:
      tags: [Customers]
      summary: 顧客を新規作成する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerCreate" }
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerResponse" }
  /customers/{id}:
    get:
      tags: [Customers]
      summary: 顧客詳細を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 顧客詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Customers]
      summary: 顧客情報を更新する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerResponse" }
...
  /consumption-tax-returns:
    get:
      tags: [ConsumptionTaxReturns]
      summary: 消費税申告一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: 消費税申告一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsumptionTaxReturnListResponse" }
    post:
      tags: [ConsumptionTaxReturns]
      summary: |
        指定した会計年度・課税方式(本則/簡易/2割特例)で消費税申告データを新規作成し計算する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fiscal_year_id, filing_method]
              properties:
                fiscal_year_id: { type: string, format: uuid }
                filing_method: { $ref: "#/components/schemas/TaxFilingMethod" }
      responses:
        "201":
          description: 作成・計算成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsumptionTaxReturnResponse" }
  /consumption-tax-returns/{id}:
    get:
      tags: [ConsumptionTaxReturns]
      summary: 消費税申告詳細(区分別内訳含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 消費税申告詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsumptionTaxReturnResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /consumption-tax-returns/{id}/calculate:
    post:
      tags: [ConsumptionTaxReturns]
      summary: 最新の仕訳データに基づき税額を再計算する(draft状態のみ)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "200":
          description: 再計算成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsumptionTaxReturnResponse" }
  /consumption-tax-returns/{id}/finalize:
    post:
      tags: [ConsumptionTaxReturns]
      summary: 消費税申告を確定する(以後は追記専用)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "200":
          description: 確定成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsumptionTaxReturnResponse" }
        "409": { $ref: "#/components/responses/Conflict" }

  # ----------------------------------------------------------------------------
  # Payroll Imports (領域D)
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWTアクセストークン。`tenant_id` クレームを含み、`X-Tenant-ID` ヘッダーと
        一致しない場合はAPI層で403 (`TENANT_MISMATCH`) を返す。
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: 操作対象テナントのUUID。JWTクレームの tenant_id と一致する必要がある。
      schema: { type: string, format: uuid }
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        副作用を伴うPOST操作の冪等性を保証するクライアント生成キー(推奨: UUID)。
        同一キーでの再送は最初のレスポンスをそのまま返す。
      schema: { type: string }
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
  responses:
...
    NotFound:
      description: リソースが見つかりません(RLSにより越境アクセス時も同一の404を返す)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: 状態遷移ルール・DB制約違反等によりリクエストを実行できません
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        pagination:
          type: object
          properties:
            page: { type: integer }
            page_size: { type: integer }
            total_count: { type: integer }
            total_pages: { type: integer }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            例: INVALID_BALANCE, VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN,
            TENANT_MISMATCH, IDEMPOTENCY_KEY_CONFLICT, VOID_WINDOW_EXPIRED,
            SELF_APPROVAL_NOT_ALLOWED, DUPLICATE_IMPORT, CONFLICT, ALREADY_EXISTS,
            INVITATION_EXPIRED, INVITATION_REVOKED, INVITATION_ALREADY_ACCEPTED
        message: { type: string }
        details:
          type: array
          items: { type: object }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: "#/components/schemas/ApiError" }

    # ---- Enums (DBのENUM型と対応) ----
    AccountType:
      type: string
      enum: [asset, liability, equity, revenue, expense]
    DebitCredit:
      type: string
      enum: [debit, credit]
...
    TaxFilingMethod:
      type: string
      enum: [general, simplified, twenty_percent_special]
...
    Account:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        code: { type: string }
        name: { type: string }
        category_id: { type: string, format: uuid, nullable: true }
        account_type: { $ref: "#/components/schemas/AccountType" }
        normal_balance: { $ref: "#/components/schemas/DebitCredit" }
        parent_account_id: { type: string, format: uuid, nullable: true }
        default_tax_category_code: { type: string, nullable: true }
        allow_manual_entry: { type: boolean, default: true }
        is_active: { type: boolean, default: true }
...
    AccountUpdate:
      type: object
      properties:
        name: { type: string }
        category_id: { type: string, format: uuid }
        allow_manual_entry: { type: boolean }
        is_active: { type: boolean }
    AccountResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/Account" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    Department:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        code: { type: string }
        name: { type: string }
        is_active: { type: boolean, default: true }
    DepartmentCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        is_active: { type: boolean, default: true }
    DepartmentResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/Department" }
        meta: { $ref: "#/components/schemas/Meta" }
    DepartmentListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/Department" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Journal Entries ----
...
    BankAccount:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        bank_name: { type: string }
        bank_code: { type: string, nullable: true }
        branch_name: { type: string, nullable: true }
        branch_code: { type: string, nullable: true }
        account_type: { type: string, enum: [ordinary, checking] }
        account_number: { type: string }
        account_holder_kana: { type: string, nullable: true }
        currency_code: { type: string, default: JPY }
        opening_balance: { type: number, default: 0 }
        linked_account_id: { type: string, format: uuid, nullable: true }
        is_active: { type: boolean, default: true }
    BankAccountCreate:
      type: object
      required: [bank_name, account_type, account_number]
      properties:
        bank_name: { type: string }
        bank_code: { type: string }
        branch_name: { type: string }
        branch_code: { type: string }
        account_type: { type: string, enum: [ordinary, checking] }
        account_number: { type: string }
        account_holder_kana: { type: string }
        currency_code: { type: string, default: JPY }
        opening_balance: { type: number, default: 0 }
        linked_account_id: { type: string, format: uuid }
    BankAccountResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/BankAccount" }
        meta: { $ref: "#/components/schemas/Meta" }
    BankAccountListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/BankAccount" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        code: { type: string }
        name: { type: string }
        kana_name: { type: string, nullable: true }
        is_active: { type: boolean, default: true }
    CustomerCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        kana_name: { type: string }
    CustomerResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/Customer" }
        meta: { $ref: "#/components/schemas/Meta" }
    CustomerListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/Customer" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    ConsumptionTaxReturnLine:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        category: { type: string, description: "例: 課税売上10%, 課税仕入8%軽減" }
        amount: { type: number }
    ConsumptionTaxReturn:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        fiscal_year_id: { type: string, format: uuid }
        filing_method: { $ref: "#/components/schemas/TaxFilingMethod" }
        taxable_sales_amount: { type: number, readOnly: true }
        taxable_purchase_amount: { type: number, readOnly: true }
        tax_due_amount: { type: number, readOnly: true }
        status: { type: string, enum: [draft, finalized] }
        finalized_at: { type: string, format: date-time, nullable: true, readOnly: true }
        lines:
          type: array
          items: { $ref: "#/components/schemas/ConsumptionTaxReturnLine" }
    ConsumptionTaxReturnResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/ConsumptionTaxReturn" }
        meta: { $ref: "#/components/schemas/Meta" }
    ConsumptionTaxReturnListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/ConsumptionTaxReturn" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Payroll Imports ----
...